+++ lib/app/config/constants.dart
@@ -20,3 +20,7 @@ const downloadsDir = 'downloads';
 const audiobooksSubDir = 'audiobooks';
 const podcastsSubDir = 'podcasts';
 const coverName = 'cover.jpg';
+
+// special symbols
+const kDot = '•';
+const kMinus = '\u2212';