+++ lib/features/player/logic/sessions_cleanup.g.dart
@@ -41,7 +41,7 @@ final class SessionsCleanupProvider
   }
 }
 
-String _$sessionsCleanupHash() => r'9c77f47038235c4fa038c475505527daf7ffbc7f';
+String _$sessionsCleanupHash() => r'13d160bfe38355d461831005d5a90b20ca104c11';
 
 abstract class _$SessionsCleanup extends $Notifier<void> {
   void build();