+++ lib/features/settings/ui/player_settings/player_settings_screen.dart
@@ -6,6 +6,7 @@ import 'package:storii/app/init.dart';
 import 'package:storii/features/settings/ui/player_settings/audio_buffer_tile.dart';
 import 'package:storii/features/settings/ui/player_settings/fade_on_sleep_tile.dart';
 import 'package:storii/features/settings/ui/player_settings/interrupt_skip_tile.dart';
+import 'package:storii/features/settings/ui/player_settings/miniplayer_subtitle_mode_tile.dart';
 import 'package:storii/features/settings/ui/player_settings/os_notification_button_tiles.dart';
 import 'package:storii/features/settings/ui/player_settings/playback_controls_layout_tile.dart';
 import 'package:storii/features/settings/ui/player_settings/player_background_theme_tile.dart';
@@ -48,22 +49,40 @@ class PlayerSettingsScreen extends ConsumerWidget {
       ),
       body: ListView(
         children: [
-          const SkipForwardTile(),
-          const SkipBackwardTile(),
-          const InterruptSkipTile(),
-          const SyncIntervalTile(),
-          const SpeedTile(),
-          const AudioBufferTile(),
-          const FadeOnSleepTile(),
-          SettingsHeader(title: l10n.miniPlayer),
-          const ShowSeekButtonsTile(),
-          SettingsHeader(title: l10n.fullPlayer),
-          const PlayerBackgroundThemeTile(),
-          const PlaybackControlsLayoutTile(),
-          SettingsHeader(title: l10n.osNotification),
-          const OsNotificationCanSeekTile(),
-          const OsNotificationCanSkipTile(),
-          const OsNotificationCanSkipChapterTile(),
+          SettingsHeader(
+            title: l10n.playback,
+            children: const [
+              SkipForwardTile(),
+              SkipBackwardTile(),
+              InterruptSkipTile(),
+              SyncIntervalTile(),
+              SpeedTile(),
+              AudioBufferTile(),
+              FadeOnSleepTile(),
+            ],
+          ),
+          SettingsHeader(
+            title: l10n.miniPlayer,
+            children: const [
+              ShowSeekButtonsTile(),
+              MiniplayerSubtitleModeTile(),
+            ],
+          ),
+          SettingsHeader(
+            title: l10n.fullPlayer,
+            children: const [
+              PlayerBackgroundThemeTile(),
+              PlaybackControlsLayoutTile(),
+            ],
+          ),
+          SettingsHeader(
+            title: l10n.osNotification,
+            children: const [
+              OsNotificationCanSeekTile(),
+              OsNotificationCanSkipTile(),
+              OsNotificationCanSkipChapterTile(),
+            ],
+          ),
         ],
       ),
     );